services/VGT/requirements.txt
fastapi==0.111.1
python-multipart==0.0.9
uvicorn==0.30.3
gunicorn==22.0.0
requests==2.32.3
torch==2.4.0
torchvision==0.19.0
timm==1.0.8
Pillow==10.4.0
pdf-annotate==0.12.0
requests==2.32.3
scipy==1.14.0
opencv-python==4.10.0.84
Shapely==2.0.5
transformers==4.40.2
pdf2image==1.17.0
lxml==5.2.2
lightgbm==4.5.0
huggingface_hub==0.24.3
setuptools==72.1.0
roman==4.2
hydra-core==1.3.2
modal
